[fdk] К предыдущему

+++ platform/fdk/src/main/java/fdk/region/ru/masterdata/Exchange.java
@@ -1,14 +1,15 @@
 package fdk.region.ru.masterdata;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 public class Exchange {
     String currencyID;
     String homeCurrencyID;
     Date date;
-    Double exchangeRate;
+    BigDecimal exchangeRate;
 
-    public Exchange(String currencyID, String homeCurrencyID, Date date, Double exchangeRate) {
+    public Exchange(String currencyID, String homeCurrencyID, Date date, BigDecimal exchangeRate) {
         this.currencyID = currencyID;
         this.homeCurrencyID = homeCurrencyID;
         this.date = date;
